fix(providers): apply Gemini review patches for TON, TRON, Bitcoin, Cosmos
- TON: BigInt precision in generateTipTransaction deep link
- Bitcoin: remove dead `sats` variable in generateTipTransaction
- TRON: TRC-20 warning in getTransaction, bandwidth/energy unit fix, BigNumber precision in tip deep link
- Cosmos: add deepLink field to generateTipTransaction, parse transfer events in cosmTxToInfo (fixed Gemini's tx.tx.body.messages — IndexedTx.tx is Uint8Array, not decoded object)
- README: merge Gemini style + full technical docs, remove duplicate section
- .gitignore: add .claude/

+++ src/providers/ton/provider.ts
@@ -194,7 +194,7 @@ export class TonProvider implements IBlockchainProvider {
     const addr = this.parseAddress(toAddress);
 
     // ton:// deep link format
-    const deepLink = `ton://transfer/${addr.toString()}?amount=${Math.round(parseFloat(amount) * 1e9)}`;
+    const deepLink = `ton://transfer/${addr.toString()}?amount=${BigInt(Math.round(parseFloat(amount) * 1e9))}`;
 
     return {
       chain: 'TON',
